requirements.txt
python-fasthtml==0.12.27
sqlite-minutils>=4.0.3